ci: add custom sync command

+++ mirror.sh
@@ -9,4 +9,10 @@ export GIT_SSH_COMMAND="ssh -v -i ~/.ssh/id_rsa -o StrictHostKeyChecking=no -l $
 git remote remove mirror || true
 
 git remote add mirror "$INPUT_TARGET_REPO_URL"
-git push --tags --force --prune mirror "refs/remotes/origin/*:refs/heads/*"
+
+if [ -z "${INPUT_SYNC_COMMAND}" ]; then
+    git push --tags --force --prune mirror "refs/remotes/origin/*:refs/heads/*"
+else
+    $INPUT_SYNC_COMMAND
+fi
+
